fix(cmd): fix cmd hierarchy to avoid cyclic deps
The way that it was before, the init() function of the sub-commands
wasn't always guaranteed to run before the root command which would
cause issues with the command processor. When corrected to be bottom up
as per the Cobra documentation, it meant that a lot of the commands
support funcs needed to be abstracted to a new file to avoid cyclic
dependencies.